+++ tests/Spectron.Emulator.Tests/Devices/UlaPlusTests.cs
@@ -54,7 +54,7 @@ public class UlaPlusTests
 
         var value = ulaPlus.ReadPort(DataPort);
 
-        value.ShouldBe((byte)0xAB);
+        value.ShouldBe(0xAB);
     }
 
     // Standard Spectrum Palette for ULA+ (G..R..B.)
